bigquery/requirements.txt
google-cloud-bigquery
pandas
db-dtypes
mysql-connector-python
google-cloud-storage